Add component for QRCode
